--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -61,37 +61,11 @@ jobs:
       - name: Checkout source
         uses: actions/checkout@v2
 
-      - name: Get yarn cache directory path
-        id: yarn-cache-dir-path
-        run: echo "::set-output name=dir::$(yarn cache dir)"
-
-      # Retain Node modules across builds; some packages download/build binaries
-      # here.
-      - name: Restore Node modules cache
-        if: runner.os != 'Windows'
-        uses: actions/cache@v2
-        id: node-modules-cache
+      - uses: actions/setup-node@v2
         with:
-          path: |
-            node_modules
-            packages/*/node_modules
-            packages/*/.vscode-test
-          key: ${{ runner.os }}-${{ matrix.node-version }}-node-modules-1-${{ hashFiles('yarn.lock')}}
-
-      - name: Restore Node modules cache for Windows
-        if: runner.os == 'Windows'
-        uses: actions/cache@v2
-        id: node-modules-cache-windows
-        with:
-          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
-          key: ${{ runner.os }}-${{ matrix.node-version }}-yarn-${{ hashFiles('**/yarn.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-${{ matrix.node-version }}-yarn-
-
-      - name: Set up Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v1
-        with:
-          node-version: ${{ matrix.node-version }}
+          node-version: "${{ matrix.node-version }}"
+          cache: "yarn"
+          cache-dependency-path: yarn.lock
 
       # Publish tests won't work if we haven't configured our committer details.
       - name: Configure Git user
@@ -99,8 +73,27 @@ jobs:
           git config --global user.name CI
           git config --global user.email ci@dendron.so
 
-      - name: Initialize
-        run: yarn ci:init
+      - name: Restore typescript lib cache
+        uses: actions/cache@v2
+        id: ts-cache
+        with:
+          path: |
+            packages/*/lib/*
+          key: ${{ runner.os }}-${{ hashFiles('yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
+
+      - name: Sets env vars for publish test
+        run: |
+          echo "TEST_NEXT_TEMPLATE=1" >> $GITHUB_ENV
+        if: runner.os == 'Linux'
+
+      - name: Bootstrap
+        run: yarn bootstrap:bootstrap
+        shell: bash
+
+      - name: Build
+        run: yarn bootstrap:buildCI
         shell: bash
 
       - name: Start Xvfb