--- a/.github/workflows/proto.yml
+++ b/.github/workflows/proto.yml
@@ -4,7 +4,10 @@
 # release.
 name: Proto
 
-on: workflow_dispatch
+on:
+  push:
+    branches:
+      - proto/*
 jobs:
   test:
     strategy: