--- a/.github/workflows/proto.yml
+++ b/.github/workflows/proto.yml
@@ -108,7 +108,7 @@ jobs:
         shell: bash
 
       - name: Build
-        run: yarn bootstrap:build
+        run: yarn bootstrap:buildCI
         shell: bash
 
       # - name: Start Xvfb