--- a/.github/workflows/proto.yml
+++ b/.github/workflows/proto.yml
@@ -55,37 +55,11 @@ jobs:
       - name: Checkout source
         uses: actions/checkout@v2
 
-      - name: Get yarn cache directory path
-        id: yarn-cache-dir-path
-        run: echo "::set-output name=dir::$(yarn cache dir)"
-
-      # Retain Node modules across builds; some packages download/build binaries
-      # here.
-      - name: Restore Node modules cache
-        if: runner.os != 'Windows'
-        uses: actions/cache@v2
-        id: node-modules-cache
+      - uses: actions/setup-node@v2
         with:
-          path: |
-            node_modules
-            packages/*/node_modules
-            packages/*/.vscode-test
-          key: ${{ runner.os }}-${{ matrix.node-version }}-node-modules-1-${{ hashFiles('yarn.lock')}}
-
-      - name: Restore Node modules cache for Windows
-        if: runner.os == 'Windows'
-        uses: actions/cache@v2
-        id: node-modules-cache-windows
-        with:
-          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
-          key: ${{ runner.os }}-${{ matrix.node-version }}-yarn-${{ hashFiles('**/yarn.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-${{ matrix.node-version }}-yarn-
-
-      - name: Set up Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v2
-        with:
-          node-version: ${{ matrix.node-version }}
+          node-version: "${{ matrix.node-version }}"
+          cache: "yarn"
+          cache-dependency-path: yarn.lock
 
       # Publish tests won't work if we haven't configured our committer details.
       - name: Configure Git user
@@ -129,6 +103,6 @@ jobs:
       #     unzip dendron-site.zip -d packages/nextjs-template
       #     pwd
 
-      - name: Run tests
-        run: yarn ci:test:${{ matrix.suite }}
-        timeout-minutes: 30
+      # - name: Run tests
+      #   run: yarn ci:test:${{ matrix.suite }}
+      #   timeout-minutes: 30