--- a/.github/workflows/proto.yml
+++ b/.github/workflows/proto.yml
@@ -25,7 +25,7 @@ jobs:
           # - 14.x
         suite:
           - cli
-          - plugin
+          # - plugin
 
         include:
           # - os: macos-latest
@@ -84,24 +84,6 @@ jobs:
       - name: Build
         run: yarn bootstrap:buildCI
         shell: bash
-
-      # - name: Start Xvfb
-      #   if: runner.os == 'Linux' && matrix.suite == 'plugin'
-      #   run: |
-      #     Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
-      #     echo 'DISPLAY=:99.0' >>$GITHUB_ENV
-
-      # - name: Download notes.zip
-      #   if: runner.os == 'Linux'
-      #   uses: wei/curl@master
-      #   with:
-      #     args: -LO https://artifacts-prod-artifactb7980f61-19orqnnuurvwy.s3.us-west-2.amazonaws.com/artifacts/dendron-site.zip
-
-      # - name: Unzip notes.zip
-      #   if: runner.os == 'Linux'
-      #   run: |
-      #     unzip dendron-site.zip -d packages/nextjs-template
-      #     pwd
 
       # - name: Run tests
       #   run: yarn ci:test:${{ matrix.suite }}