--- a/.github/workflows/proto.yml
+++ b/.github/workflows/proto.yml
@@ -83,6 +83,12 @@ jobs:
         run: yarn bootstrap:buildCI
         shell: bash
 
+      - name: Start Xvfb
+        if: runner.os == 'Linux' && matrix.suite == 'plugin'
+        run: |
+          Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
+          echo 'DISPLAY=:99.0' >>$GITHUB_ENV
+
       - name: Run tests
         run: yarn ci:test:${{ matrix.suite }}
       #   timeout-minutes: 30