--- a/bootstrap/scripts/buildAllForTest.js
+++ b/bootstrap/scripts/buildAllForTest.js
@@ -11,17 +11,36 @@ const $ = (cmd) => {
   return execa.commandSync(cmd, { stdout: process.stdout, buffer: false });
 };
 
+const TEST_NEXT_TEMPLATE = process.env.TEST_NEXT_TEMPLATE;
+
 console.log("build all...");
 $(`npx lerna run buildCI --scope @dendronhq/common-all`);
 $(`npx lerna run buildCI --scope @dendronhq/common-server `);
 $(`npx lerna run buildCI --scope @dendronhq/engine-server `);
 $(`npx lerna run buildCI --scope @dendronhq/pods-core `);
-$(
-  `npx lerna run buildCI --parallel --scope "@dendronhq/{common-test-utils,api-server,common-assets}"`
-);
+if (TEST_NEXT_TEMPLATE) {
+  $(
+    `npx lerna run buildCI --parallel --scope "@dendronhq/{common-test-utils,api-server,common-assets}"`
+  );
+} else {
+  $(
+    `npx lerna run buildCI --parallel --scope "@dendronhq/{common-test-utils,api-server}"`
+  );
+}
 $(
   `npx lerna run buildCI --parallel --scope "@dendronhq/{common-frontend,dendron-cli}"`
 );
-$(`npx lerna run buildCI --scope "@dendronhq/engine-test-utils" `);
+if (TEST_NEXT_TEMPLATE) {
+  $(
+    `npx lerna run buildCI --scope "@dendronhq/{engine-test-utils,dendron-next-server}" `
+  );
+} else {
+  $(`npx lerna run buildCI --scope "@dendronhq/engine-test-utils" `);
+}
+
 $(`npx lerna run buildCI --scope "@dendronhq/plugin-core"`);
+
+if (TEST_NEXT_TEMPLATE) {
+  $(`npx yarn dendron dev sync_assets --fast`);
+}
 console.log("done");