… package.json
… packages/api-server/package.json
… packages/common-all/package.json
… packages/common-assets/package.json
… packages/common-frontend/package.json
… packages/common-server/package.json
… packages/common-test-utils/package.json
… packages/dendron-cli/package.json
… packages/dendron-next-server/package.json
… packages/dendron-plugin-views/package.json
… packages/engine-server/package.json
… packages/engine-test-utils/package.json
… packages/lsp-client/package.json
… packages/plugin-core/package.json
… packages/pods-core/package.json